Fix technique filename not being expandable if it appears in the list multiple times due to reordering (#4)
If the user reorders techniques by drag/drop and drags a technique above another technique so the filename the technique is in appears multiple times in the techniques tree, the first node with the technique filename was expandable manually, but all other nodes were not because the ID of the label was the same. (They also expanded when the first appearance node of the filename was expanded manually). This fixes that.