Increase compression effort for setup/installer.
sets the compression effort for the zip archive attached to the installer up from the default to almost the absolute maximum. 
In the future if the version of .NET used is updated to a modern version of .NET Core then add `-mm=Deflate64` and replace `-mfb=258` with `-mfb=257` for a an even further improved compression ratio.