Rework OpenVR hooks to hook "openvr_api.dll" instead of "vrclient.dll" again
The SteamVR client library may be loaded and unloaded (as OpenXR runtime vs. for OpenVR), which is prevented by ReShade pinning hooked modules.
This breaks the library, as it does not support reinitializing in a different mode (and some games first load OpenXR and afterwards switch to OpenVR).
To solve this, do not hook the SteamVR client library at all, instead only hook the OpenVR loader that sits before it.