Add a custom fixed RIID/GUID to access the parent proxy object (or well, child, depending on the point of view) to facilitate applications trying to work around the multiple layers of proxies that we are getting into due to: Overlays, SpecialK, ReShade, Nvidia Streamline, AMD FidelityFX, other custom hooks, etc etc As of now there's no good way to go up or down the chain.
Additionally, this can be good for ReShade addons that directly work at device level (without passing through the ReShade API) to create resources, and they don't want, or need, ReShade to keep track of the resources they create (the idea would be for these to be confined to the ones originally created by the application).

For now this is just implemented in DX11/DXGI, as I don't use DX10 and 12 and thus can't test it. DX9 uses the same system too so it's a candidate.
The idea would be to push the same exact GUID to all the applications listed above, so we can all work around their limitations.